Move GuessResult to core constants
