+++ battleship-core/src/lib.rs
@@ -6,7 +6,6 @@ pub mod ship;
 pub use battleship_common::BoardView;
 pub use battleship_config::{GRID_SIZE, SHIPS};
 pub use board::Board;
-pub use constants::{Cell, GameplayError, GuessError, PlayerState};
-pub use battleship_common::GuessResult as GuessResult;
+pub use constants::{Cell, GameplayError, GuessError, GuessResult, PlayerState};
 pub use fleet::Fleet;
 pub use ship::Ship;
